+++ test/better_mailer_previews_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class BetterMailerPreviewsTest < ActiveSupport::TestCase
-  test "it has a version number" do
-    assert BetterMailerPreviews::VERSION
-  end
-end